refactor: migrate fix_intent, fix_plan_dismissals, policy_compiler, repair_template_registry, incremental to drift-engine
(cherry picked from commit 0cac4eb199a4c7fef8f3e91b95ff5d6bdd21e36d)
